Added Turns, Updated DisplayManager, Added ArmyManager
Display Manager now displays which turn it is becoming.

Added Army Manager that reads the army json and populates the units.

Updated Tiles for Generations to match how I did it with the Player Army json generation. Felt like a better system and would leave the tiles in the game to protect their values.

Once all units in an army have acted, their turn ends.